feat: edit backend files

+++ backend/controllers/cards.js
@@ -1,45 +1,51 @@
 const Card = require("../models/cards");
-const { checkError } = require("../utils/errors");
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
-    .then((user) => res.send({ data: user }))
-    .catch((err) => checkError(res, err));
+    .then((cards) => res.send(cards))
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.send({ data: card }))
-    .catch((err) => checkError(res, err));
+    .then((card) => res.status(201).send(card))
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new Error("notValidId"))
-    .then((card) => res.send({ data: card }))
-    .catch((err) => checkError(res, err));
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(new Error('cardNotFound'))
+    .then((data) => {
+      // eslint-disable-next-line eqeqeq
+      if (data.owner != req.user._id) {
+        throw new Error('forbidden');
+      }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((card) => res.send(card));
+    })
+    .catch(next);
 };
 
-module.exports.putLike = (req, res) => {
+module.exports.putLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error("notValidId"))
-    .then((card) => res.send({ data: card }))
-    .catch((err) => checkError(res, err));
+    .orFail(new Error('cardNotFound'))
+    .then((card) => res.send(card))
+    .catch(next);
 };
 
-module.exports.deleteLike = (req, res) => {
+module.exports.deleteLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error("notValidId"))
-    .then((card) => res.send({ data: card }))
-    .catch((err) => checkError(res, err));
+    .orFail(new Error('cardNotFound'))
+    .then((card) => res.send(card))
+    .catch(next);
 };
